claim-schema changes per meeting
Several changes to claim-schema.json were brought up during the team meeting
earlier.  As such, the changes were made, and a corresponding GoLang binding
was generated.  The generator was severely improved to allow remapping
properties at an arbitrary depth.  Hitherto, this was impossible, and would
consistently break the generator.

Additionally, it was discovered that the auto-generator wasn't automatically
attributing a license to the client binding.  This was changed, and is
configurable through generated-license-header.txt.

Tests were modified to ensure this is working properly.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>