Augment claim schema to allow arbitrary "nodes" payload
CTONET-676 tracks a bug in which a claim's "nodes" data is output in a manner
inconsistent with the `claim.schema.json` JSON schema.  This PR augments the
schema, as the actual implementation was found to output correct information.

Additionally, this PR moves `claim.schema.json` to a `schemas` directory in an
attempt to stay consistent with the `test-network-function` repository layout.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>